requirements.txt
anyio==3.6.1
appnope==0.1.3
asttokens==2.2.1
autopep8==1.7.0
backcall==0.2.0
click==8.1.3
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fastapi==0.85.1
h11==0.14.0
httptools==0.5.0
idna==3.4
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
stack-data==0.6.2
starlette==0.20.4
toml==0.10.2
tornado==6.2
traitlets==5.6.0
typing_extensions==4.4.0
uvicorn==0.18.3
uvloop==0.17.0
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
